requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==7.1.2
dataclasses-json==0.6.1
dnspython==2.4.2
Flask==1.1.2
frozenlist==1.4.0
gunicorn==20.0.4
idna==3.4
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.329
langsmith==0.0.56
loguru==0.7.2
MarkupSafe==1.1.1
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==0.28.1
packaging==23.2
pinecone-client==2.2.4
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==1.0.1
yarl==1.9.2